getting ready to tag

+++ py/desiUtil/install/main.py
@@ -40,6 +40,8 @@ def main():
         help='Build any Sphinx or Doxygen documentation.')
     parser.add_argument('-F', '--force', action='store_true', dest='force',
         help='Overwrite any existing installation of this product/version.')
+    parser.add_argument('-k', '--keep', action='store_true', dest='keep',
+        help='Keep the exported build directory.')
     parser.add_argument('-m', '--module-home', action='store', dest='moduleshome',
         metavar='DIR',help='Set or override the value of $MODULESHOME',
         default=getenv('MODULESHOME'))
@@ -177,6 +179,8 @@ def main():
         else:
             logger.error("DESI_PRODUCT_ROOT is missing or not set.")
             return 1
+    if options.root is not None:
+        environ['DESI_PRODUCT_ROOT'] = options.root
     install_dir = join(options.root,baseproduct,baseversion)
     if isdir(install_dir) and not options.test:
         if options.force:
@@ -286,6 +290,9 @@ set ModulesVersion "{0}"
             #
             # Assume Sphinx documentation.
             #
+            logger.debug("Found Sphinx documentation.")
+            logger.debug("module('load','{0}/{1}')".format(baseproduct,baseversion))
+            module('load',baseproduct+'/'+baseversion)
             sphinx_keywords = {
                 'name':baseproduct,
                 'release':baseversion,
@@ -294,7 +301,7 @@ set ModulesVersion "{0}"
             for sd in ('_templates','_build','_static'):
                 if not isdir(join('doc',sd)):
                     try:
-                        makedirs(install_dir)
+                        makedirs(join('doc',sd))
                     except OSError as ose:
                         logger.error(ose.strerror)
                         return 1
@@ -302,7 +309,7 @@ set ModulesVersion "{0}"
                 copyfile(join(getenv('DESIUTIL_DIR'),'etc','doc','Makefile'),
                     join('doc','Makefile'))
             if not exists(join('doc','conf.py')):
-                with open(join(getenv('DESIUTIL_DIR'),'etc','doc','Makefile')) as conf:
+                with open(join(getenv('DESIUTIL_DIR'),'etc','doc','conf.py')) as conf:
                     newconf = conf.read().format(**sphinx_keywords)
                 with open(join('doc','conf.py'),'w') as conf2:
                     conf2.write(newconf)
@@ -316,8 +323,9 @@ set ModulesVersion "{0}"
                     logger.error("Error during documentation build:")
                     logger.error(err)
                     return 1
-            if isdir(join('build','html')):
-                copytree(join('build','html'),join(install_dir,'doc'))
+            if not options.test:
+                if isdir(join('build','sphinx','html')):
+                    copytree(join('build','sphinx','html'),join(install_dir,'doc'))
         else:
             logger.warn("Documentation build requested, but no documentation found.")
 
@@ -325,7 +333,8 @@ set ModulesVersion "{0}"
     # Clean up
     #
     chdir(original_dir)
-    rmtree(working_dir)
+    if not options.keep:
+        rmtree(working_dir)
     return 0
 #
 #
